Fixed an issue where the assignment editor's save button would appear without making changes (#725)
* Improved target points check

* Fixed format editor save button appearing bug

* Fixed linter issue

* Fixed file name

* Added max to preset node target points

* Fixed migrations

* Added frontend check for target too high

+++ src/django/VLE/models.py
@@ -817,7 +817,7 @@ class PresetNode(models.Model):
         choices=TYPES,
     )
 
-    target = models.IntegerField(
+    target = models.FloatField(
         null=True,
     )
